defined abstract Piece class, allowing to calculate legal moves
