+++ src/main/java/src/pieces/Piece.java
@@ -1,0 +1,21 @@
+package src.pieces;
+
+import src.Alliance;
+import src.board.Board;
+import src.board.Move;
+
+import java.util.List;
+
+public abstract class Piece {
+
+    protected final int piecePosition;
+    protected final Alliance pieceAlliance;
+
+    public Piece(final int piecePosition, final Alliance pieceAlliance) {
+        this.piecePosition = piecePosition;
+        this.pieceAlliance = pieceAlliance;
+    }
+
+    public abstract List<Move> calculateMoves(final Board board);
+
+}
